project12/requirements.txt
streamlit==1.32.0
fastapi==0.110.0
uvicorn==0.27.1
groq==0.4.2
python-dotenv==1.0.1
pydantic==2.6.3
python-multipart==0.0.9
websockets==12.0
redis==5.0.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-whatsapp-sdk==0.1.2
elevenlabs==0.3.0
webrtc==0.0.1
plotly==5.18.0
pandas==2.2.0
numpy==1.26.4